requirements.txt
# native dependencies
Pillow==9.3.0
lxml==4.9.1
psycopg2==2.9.5
Django==3.2.16

# Other
amqp==5.1.1
beautifulsoup4==4.11.1
httplib2<0.21.1
hyperlink==21.0.0
idna>=2.5,<3.5
urllib3==1.26.12
Paver==1.3.4
python-slugify==6.1.2
decorator==5.1.1
celery==5.2.7
kombu==5.2.4
vine==5.0.0
tqdm==4.64.1
Deprecated==1.2.13
wrapt==1.14.1
jsonschema==4.17.0
zipstream-new==1.1.8
schema==0.7.5
rdflib==6.1.1
smart_open==6.2.0
PyMuPDF==1.20.2

# Django Apps
django-allauth==0.51.0
django-appconf==1.0.5
django-celery-results==2.4.0
django-filter==22.1
django-imagekit==4.1.0
django-taggit==1.5.1
django-markdownify==0.9.2
django-mptt==0.14.0
django-modeltranslation>=0.11,<0.19.0
django-treebeard==4.5.1
django-guardian<2.4.1
django-downloadview==2.3.0
django-polymorphic==3.1.0
django-tastypie<0.15.0
django-tinymce==3.5.0
django-grappelli==3.0.3
django-uuid-upload-path==1.0.0
django-widget-tweaks==1.4.12
django-sequences==2.7
oauthlib==3.2.2
pyopenssl==22.1.0
pyjwt==2.6.0

# geopython dependencies
pyproj<3.5.0
OWSLib==0.27.2
pycsw==2.6.1
SQLAlchemy==1.4.43 # required by PyCSW
Shapely==1.8.5.post1
mercantile==1.2.1
geoip2==4.6.0
numpy==1.23.*

# # Apps with packages provided in GeoNode's PPA on Launchpad.

# Django Apps
dj-database-url==1.0.0
dj-pagination==2.5.0
django-select2==8.0.0
django-floppyforms<1.10.0
django-forms-bootstrap<=3.1.0
django-autocomplete-light==3.5.1
django-invitations<2.0.1
django-recaptcha==3.0.0

# REST
djangorestframework==3.12.0
djangorestframework-gis==1.0
djangorestframework-guardian==0.3.0
drf-extensions==0.7.1
drf-writable-nested==0.7.0
drf-spectacular==0.24.2
dynamic-rest==2.1.2
Markdown==3.4.1

pinax-notifications==6.0.0
pinax-ratings==4.0.0

# GeoNode org maintained apps.
# django-geonode-mapstore-client==4.0.4
-e git+https://github.com/sp1ritnsk/geonode-mapstore-client.git@master#egg=django_geonode_mapstore_client
geonode-avatar==5.0.8
geonode-oauth-toolkit==2.2.2
geonode-user-messages==2.0.2
geonode-announcements==2.0.2
geonode-django-activity-stream==0.10.0
gn-arcrest==10.5.5
geoserver-restconfig==2.0.8
gn-gsimporter==2.0.4
gisdata==0.5.4

# haystack/elasticsearch
django-haystack==3.2.1
elasticsearch>=2.0.0,<9.0.0

# datetimepicker widget
django-bootstrap3-datetimepicker-2==2.8.3

# storage manager dependencies
django-storages==1.13.1
dropbox==11.35.0
google-cloud-storage==2.5.0
google-cloud-core==2.3.2
boto3==1.26.4

# Django Caches
python-memcached<=1.59
whitenoise==6.2.0
Brotli==1.0.9

# Contribs
xmltodict<0.13.1
inflection>=0.4.0
jdcal==1.4.1
mock<5.0.0
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
timeout-decorator==0.5.0
pylibmc==1.6.3
sherlock==0.4.0

# required by monitoring
psutil==5.9.3
django-cors-headers==3.13.0
user-agents
django-user-agents
xmljson
django-ipware<4.1
# no version here, use latest one with fresh data
pycountry

# production
uWSGI==2.0.21
gunicorn==20.1.0
ipython==8.6.0
docker==6.0.1
invoke==1.7.3

# tests
coverage==6.5.0
requests-toolbelt==0.10.1
flake8==5.0.4
pytest==7.2.0
pytest-bdd==6.1.0
splinter==0.18.1
pytest-splinter==3.3.2
pytest-django==4.5.2
setuptools>=59.1.1,<65.6.0
pip==22.3.1
Twisted==22.10.0
pixelmatch==0.3.0
factory-boy==3.2.1
flaky==3.7.0
selenium>=4.1.0,<5.0.0
selenium-requests==2.0.0
webdriver_manager==3.8.4

# Security and audit
mistune==2.0.4
wandb==0.13.5
protobuf==3.20.3
mako==1.2.3
